move filetype_is_file to its own module
